integ-app/backend/requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
pillow
python-multipart
numpy
